refactor(canvas): 캔버스 위젯 파일 구조 및 네이밍 체계 통일
- 파일 구조 재조직화
  * drawing_tools/, editor_controls/ 폴더 → toolbar/, controls/로 변경
  * 논리적 기능별 분류: 도구 관련 vs 컨트롤 관련
  * 독립 위젯들은 widgets/ 루트에 배치

- 네이밍 컨벤션 통일
  * 모든 파일명을 note_editor_ 접두사로 통일
  * 클래스명을 NoteEditor 접두사로 통일하여 파일명과 일치시킴
  * 기능별 접미사 도입 (_selector, _toggle, _toolbar, _button 등)

- Import 경로 및 클래스 참조 업데이트
  * 모든 파일에서 변경된 클래스명으로 업데이트
  * 폴더 구조 변경에 따른 import 경로 수정

- 향후 확장성 개선
  * 새로운 위젯 추가 시 명확한 분류 기준 제공
  * 일관된 네이밍으로 코드 가독성 및 유지보수성 향상

파일 변경 사항:
- 이동: 8개 파일 (toolbar/, controls/ 폴더로)
- 클래스명 변경: 11개 클래스
- Import 업데이트: 3개 페이지 파일

lib/features/canvas/widgets/
├── toolbar/                                    # í» ️ 도구 모음
│   ├── note_editor_toolbar.dart               # 메인 툴바 (컨테이너)
│   ├── note_editor_actions_bar.dart           # 액션 바 (실행취소/다시실행)
│   ├── note_editor_drawing_toolbar.dart       # 그리기 도구 모음
│   ├── note_editor_color_selector.dart        # 색상 선택기
│   ├── note_editor_stroke_selector.dart       # 굵기 선택기
│   ├── note_editor_tool_selector.dart         # 도구 선택기 (펜/지우개/하이라이터)
│   └── note_editor_color_button.dart          # 색상 버튼 (재사용 컴포넌트)
├── controls/                                   # ⚙️ 에디터 컨트롤
│   ├── note_editor_pointer_mode.dart          # 포인터 모드 토글
│   ├── note_editor_pressure_toggle.dart       # 필압 토글
│   └── note_editor_viewport_info.dart         # 뷰포트 정보 표시
└── note_editor_background_placeholder.dart    # í¶¼️ 배경 위젯
