+++ lib/features/canvas/routing/canvas_routers.dart
@@ -1,8 +1,8 @@
 import 'package:go_router/go_router.dart';
 
-import '../../../features/notes/data/notes.dart';
+import '../../../features/notes/data/fake_notes.dart';
 import '../../../shared/routing/app_routes.dart';
-import '../pages/canvas_page.dart';
+import '../pages/note_editor_screen.dart';
 
 /// 🎨 캔버스 기능 관련 라우트 설정
 ///
@@ -18,7 +18,7 @@ class CanvasRouters {
         // TODO(추후): noteId를 사용해서 실제 노트 데이터 로드
         // 현재는 임시로 tmpNote 사용
         print('📝 노트 편집 페이지: noteId = $noteId');
-        return CanvasPage(note: tmpNote);
+        return NoteEditorScreen(note: fakeNote);
       },
     ),
   ];
