+++ lib/features/canvas/widgets/toolbar/note_editor_stroke_selector.dart
@@ -1,0 +1,67 @@
+import 'package:flutter/material.dart';
+import 'package:scribble/scribble.dart';
+
+import '../../notifiers/custom_scribble_notifier.dart';
+
+class NoteEditorStrokeSelector extends StatelessWidget {
+  const NoteEditorStrokeSelector({
+    required this.notifier,
+    super.key,
+  });
+
+  final CustomScribbleNotifier notifier;
+
+  @override
+  Widget build(BuildContext context) {
+    return ValueListenableBuilder<ScribbleState>(
+      valueListenable: notifier,
+      builder: (context, state, _) => Row(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: [
+          for (final w in notifier.toolMode.widths)
+            _buildStrokeButton(
+              context,
+              strokeWidth: w,
+              state: state,
+            ),
+        ],
+      ),
+    );
+  }
+
+  Widget _buildStrokeButton(
+    BuildContext context, {
+    required double strokeWidth,
+    required ScribbleState state,
+  }) {
+    final selected = state.selectedWidth == strokeWidth;
+    return Padding(
+      padding: const EdgeInsets.all(4),
+      child: Material(
+        elevation: selected ? 4 : 0,
+        shape: const CircleBorder(),
+        child: InkWell(
+          onTap: () => notifier.setStrokeWidth(strokeWidth),
+          customBorder: const CircleBorder(),
+          child: AnimatedContainer(
+            duration: kThemeAnimationDuration,
+            width: strokeWidth * 2,
+            height: strokeWidth * 2,
+            decoration: BoxDecoration(
+              color: state.map(
+                drawing: (s) => Color(s.selectedColor),
+                erasing: (_) => Colors.transparent,
+              ),
+              border: state.map(
+                drawing: (_) => null,
+                erasing: (_) => Border.all(width: 1),
+              ),
+              borderRadius: BorderRadius.circular(50.0),
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
